services/ai-insights-service/requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
python-dotenv
google-generativeai
passlib[bcrypt] # For JWT deps
python-jose[cryptography] # For JWT deps